Implemented part of ticket #477 - Changed the behavior of "clear TABLE()" - Renamed "tocache" with "totable" (old one still persists) - Added "TABLE().name" method - Added some more DoxyGen comments - Re-enabled the replacement of "&PATH&" in the documentation files with their actual path equivalent - Resolved an issue in filtering.hpp - Enabled use of table methods in string expressions - Improved the table method highlighting in the terminal
git-svn-id: https://svn.code.sf.net/p/numere/source/branches/ticket_477@755 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ kernel/core/maths/filtering.hpp
@@ -209,8 +209,8 @@ namespace NumeRe
                 {
                     for (size_t j = 0; j < m_windowSize.second; j++)
                     {
-                        if (sqrt(pow2(i/(max(1u, m_windowSize.first)-1)-mean_row) + pow2(j/(max(1u, m_windowSize.second)-1)-mean_col)) <= 0.5)
-                            m_filterKernel[i][j] = fabs(sqrt(pow2(i/(max(1u, m_windowSize.first)-1)-mean_row) + pow2(j/(max(1u, m_windowSize.second)-1)-mean_col)) - 0.5);
+                        if (sqrt(pow2(i/((double)max(1u, m_windowSize.first-1))-mean_row) + pow2(j/((double)max(1u, m_windowSize.second-1))-mean_col)) <= 0.5)
+                            m_filterKernel[i][j] = fabs(sqrt(pow2(i/((double)max(1u, m_windowSize.first-1))-mean_row) + pow2(j/((double)max(1u, m_windowSize.second-1))-mean_col)) - 0.5);
                     }
                 }
             }
